+++ blue_users.py
@@ -1,44 +1,68 @@
 import csv
+import json
 import os
 
-from flask import Blueprint, request, render_template, Response
+import flask_login
+from flask import Blueprint, render_template, request, Response
+from flask_login import login_required
+from werkzeug.utils import redirect
+
 from classes.Database import Database
-import json
 
 users = Blueprint('users', __name__)
-basedir = os.path.abspath(os.path.dirname(__file__))
 
 
-@users.route('/users', methods=['GET'])
-def reval_users():
-    return render_template('users.html')
+@users.route('/users/', methods=['GET'])
+@login_required
+def _users():
+    user = flask_login.current_user
+    if user.is_admin():
+        return render_template('users.html')
+    else:
+        return redirect(request.referrer)
 
 
-@users.route('/getUsers', methods=['GET'])
+@users.route('/users/get')
+@login_required
 def get_users():
     db = Database()
     db.conn.row_factory = db.dict_factory
     c = db.conn.cursor()
-    sql_string = 'SELECT * FROM users'
-    c.execute(sql_string)
-    data = c.fetchall()
+
+    user = flask_login.current_user
+
+    # admin is allowed to receive all the data
+    if user.is_admin():
+        sql_string = 'SELECT * FROM users '
+        c.execute(sql_string)
+        data = c.fetchall()
+    # the rest is only allowed to see the insensitive data
+    else:
+        data = []
 
     c.close()
     return json.dumps(data)
 
-@users.route('/updateUser', methods=['GET'])
+
+@users.route('/users/update', methods=['GET'])
+@login_required
 def update_user():
+    user = flask_login.current_user
+    # only admin is allowed to update a user in the user table
+    if not user.is_admin():
+        return 'http403'
+
     sql_string = 'UPDATE users SET '
     sql_parameters = ''
     ID = request.args.get('ID')
 
-    user_last_name = request.args.get('user_last_name')
+    user_last_name = request.args.get('user_last_name').upper()
     if user_last_name != '':
         sql_parameters = sql_parameters + 'user_last_name="' + user_last_name + '"'
     else:
         return 'http400'
 
-    user_name = request.args.get('user_name')
+    user_name = request.args.get('user_name').title()
     if user_name != '':
         sql_parameters = sql_parameters + ',user_name="' + user_name + '"'
     else:
@@ -47,14 +71,6 @@ def update_user():
     user_sex = request.args.get('user_sex')
     if user_sex != '':
         sql_parameters = sql_parameters + ',user_sex="' + user_sex + '"'
-
-    user_is_pi = request.args.get('user_is_pi')
-    if user_is_pi != '':
-        sql_parameters = sql_parameters + ',user_is_pi="' + user_is_pi + '"'
-
-    user_is_phd = request.args.get('user_is_phd')
-    if user_is_phd != '':
-        sql_parameters = sql_parameters + ',user_is_phd="' + user_is_phd + '"'
 
     user_title = request.args.get('user_title')
     if user_title != '':
@@ -71,6 +87,8 @@ def update_user():
     user_email = request.args.get('user_email')
     if user_email != '':
         sql_parameters = sql_parameters + ',user_email="' + user_email + '"'
+    else:
+        return 'http400'
 
     user_home_address = request.args.get('user_home_address')
     if user_home_address != '':
@@ -79,6 +97,8 @@ def update_user():
     user_telephone = request.args.get('user_telephone')
     if user_telephone != '':
         sql_parameters = sql_parameters + ',user_telephone="' + user_telephone + '"'
+    else:
+        return 'http400'
 
     user_private_phone = request.args.get('user_private_phone')
     if user_private_phone != '':
@@ -92,41 +112,39 @@ def update_user():
     if user_out_date != '':
         sql_parameters = sql_parameters + ',user_out_date="' + user_out_date + '"'
 
-    user_pw_hash = request.args.get('user_pw_hash')
-    if user_pw_hash != '':
-        sql_parameters = sql_parameters + ',user_pw_hash="' + user_pw_hash + '"'
+    user_is_pi = request.args.get('user_is_pi')
+    if user_is_pi != '':
+        sql_parameters = sql_parameters + ',user_is_pi="' + user_is_pi + '"'
 
-    user_alternative_ID = request.args.get('user_alternative_ID')
-    if user_alternative_ID != '':
-        sql_parameters = sql_parameters + ',user_alternative_ID="' + user_alternative_ID + '"'
+    user_is_phd = request.args.get('user_is_phd')
+    if user_is_phd != '':
+        sql_parameters = sql_parameters + ',user_is_phd="' + user_is_phd + '"'
 
-    user_can_see_private_data = request.args.get('user_can_see_private_data')
-    if user_can_see_private_data != '':
-        sql_parameters = sql_parameters + ',user_can_see_private_data="' + user_can_see_private_data + '"'
+    user_is_financial_team = request.args.get('user_is_financial_team')
+    if user_is_financial_team != '':
+        sql_parameters = sql_parameters + ',user_is_financial_team="' + user_is_financial_team + '"'
 
-    user_can_see_financial_data = request.args.get('user_can_see_financial_data')
-    if user_can_see_financial_data != '':
-        sql_parameters = sql_parameters + ',user_can_see_financial_data="' + user_can_see_financial_data + '"'
+    user_is_admin = request.args.get('user_is_admin')
+    if user_is_admin != '':
+        sql_parameters = sql_parameters + ',user_is_admin="' + user_is_admin + '"'
 
-    user_is_users_admin = request.args.get('user_is_users_admin')
-    if user_is_users_admin != '':
-        sql_parameters = sql_parameters + ',user_is_users_admin="' + user_is_users_admin + '"'
+    user_is_lender = request.args.get('user_is_lender')
+    if user_is_lender != '':
+        sql_parameters = sql_parameters + ',user_is_lender="' + user_is_lender + '"'
 
-    user_is_equipment_admin = request.args.get('user_is_equipment_admin')
-    if user_is_equipment_admin != '':
-        sql_parameters = sql_parameters + ',user_is_equipment_admin="' + user_is_equipment_admin + '"'
+    user_is_lender_admin = request.args.get('user_is_lender_admin')
+    if user_is_lender_admin != '':
+        sql_parameters = sql_parameters + ',user_is_lender_admin="' + user_is_lender_admin + '"'
 
-    user_is_suppliers_admin = request.args.get('user_is_suppliers_admin')
-    if user_is_suppliers_admin != '':
-        sql_parameters = sql_parameters + ',user_is_suppliers_admin="' + user_is_suppliers_admin + '"'
+    user_is_owner = request.args.get('user_is_owner')
+    if user_is_owner != '':
+        sql_parameters = sql_parameters + ',user_is_owner="' + user_is_owner + '"'
 
     db = Database()
-    print(sql_parameters)
 
     if sql_parameters[0] == ',':
         sql_parameters = sql_parameters[1:]
     sql_string = sql_string + sql_parameters + ' WHERE ID = ' + ID
-    print(sql_string)
 
     c = db.conn.cursor()
     c.execute(sql_string)
@@ -135,20 +153,27 @@ def update_user():
     c.close()
     return 'http200'
 
-@users.route('/newUser', methods=['GET'])
+
+@users.route('/users/new', methods=['GET'])
+@login_required
 def new_user():
+    user = flask_login.current_user
+    # only admin is allowed to create a new user in the user table
+    if not user.is_admin():
+        return 'http403'
+
     sql_string = 'INSERT INTO users '
     sql_parameters = '('
     sql_values = '('
 
-    user_last_name = request.args.get('user_last_name')
+    user_last_name = request.args.get('user_last_name').upper()
     if user_last_name != '':
         sql_parameters = sql_parameters + 'user_last_name,'
         sql_values = sql_values + '"' + user_last_name + '",'
     else:
         return 'http400'
 
-    user_name = request.args.get('user_name')
+    user_name = request.args.get('user_name').title()
     if user_name != '':
         sql_parameters = sql_parameters + 'user_name,'
         sql_values = sql_values + '"' + user_name + '",'
@@ -159,16 +184,6 @@ def new_user():
     if user_sex != '':
         sql_parameters = sql_parameters + 'user_sex,'
         sql_values = sql_values + '"' + user_sex + '",'
-
-    user_is_pi = request.args.get('user_is_pi')
-    if user_is_pi != '':
-        sql_parameters = sql_parameters + 'user_is_pi,'
-        sql_values = sql_values + '"' + user_is_pi + '",'
-
-    user_is_phd = request.args.get('user_is_phd')
-    if user_is_phd != '':
-        sql_parameters = sql_parameters + 'user_is_phd,'
-        sql_values = sql_values + '"' + user_is_phd + '",'
 
     user_title = request.args.get('user_title')
     if user_title != '':
@@ -189,6 +204,8 @@ def new_user():
     if user_email != '':
         sql_parameters = sql_parameters + 'user_email,'
         sql_values = sql_values + '"' + user_email + '",'
+    else:
+        return 'http400'
 
     user_home_address = request.args.get('user_home_address')
     if user_home_address != '':
@@ -199,6 +216,8 @@ def new_user():
     if user_telephone != '':
         sql_parameters = sql_parameters + 'user_telephone,'
         sql_values = sql_values + '"' + user_telephone + '",'
+    else:
+        return 'http400'
 
     user_private_phone = request.args.get('user_private_phone')
     if user_private_phone != '':
@@ -215,47 +234,47 @@ def new_user():
         sql_parameters = sql_parameters + 'user_out_date,'
         sql_values = sql_values + '"' + user_out_date + '",'
 
-    user_pw_hash = request.args.get('user_pw_hash')
-    if user_pw_hash != '':
-        sql_parameters = sql_parameters + 'user_pw_hash,'
-        sql_values = sql_values + '"' + user_pw_hash + '",'
+    user_is_pi = request.args.get('user_is_pi')
+    if user_is_pi != '':
+        sql_parameters = sql_parameters + 'user_is_pi,'
+        sql_values = sql_values + '"' + user_is_pi + '",'
 
-    user_alternative_ID = request.args.get('user_alternative_ID')
-    if user_alternative_ID != '':
-        sql_parameters = sql_parameters + 'user_alternative_ID,'
-        sql_values = sql_values + '"' + user_alternative_ID + '",'
+    user_is_phd = request.args.get('user_is_phd')
+    if user_is_phd != '':
+        sql_parameters = sql_parameters + 'user_is_phd,'
+        sql_values = sql_values + '"' + user_is_phd + '",'
 
-    user_can_see_private_data = request.args.get('user_can_see_private_data')
-    if user_can_see_private_data != '':
-        sql_parameters = sql_parameters + 'user_can_see_private_data,'
-        sql_values = sql_values + '"' + user_can_see_private_data + '",'
+    user_is_financial_team = request.args.get('user_is_financial_team')
+    if user_is_financial_team != '':
+        sql_parameters = sql_parameters + 'user_is_financial_team,'
+        sql_values = sql_values + '"' + user_is_financial_team + '",'
 
-    user_can_see_financial_data = request.args.get('user_can_see_financial_data')
-    if user_can_see_financial_data != '':
-        sql_parameters = sql_parameters + 'user_can_see_financial_data,'
-        sql_values = sql_values + '"' + user_can_see_financial_data + '",'
+    user_is_admin = request.args.get('user_is_admin')
+    if user_is_admin != '':
+        sql_parameters = sql_parameters + 'user_is_admin,'
+        sql_values = sql_values + '"' + user_is_admin + '",'
 
-    user_is_users_admin = request.args.get('user_is_users_admin')
-    if user_is_users_admin != '':
-        sql_parameters = sql_parameters + 'user_is_users_admin,'
-        sql_values = sql_values + '"' + user_is_users_admin + '",'
+    user_is_lender = request.args.get('user_is_lender')
+    if user_is_lender != '':
+        sql_parameters = sql_parameters + 'user_is_lender,'
+        sql_values = sql_values + '"' + user_is_lender + '",'
 
-    user_is_equipment_admin = request.args.get('user_is_equipment_admin')
-    if user_is_equipment_admin != '':
-        sql_parameters = sql_parameters + 'user_is_equipment_admin,'
-        sql_values = sql_values + '"' + user_is_equipment_admin + '",'
+    user_is_lender_admin = request.args.get('user_is_lender_admin')
+    if user_is_lender_admin != '':
+        sql_parameters = sql_parameters + 'user_is_lender_admin,'
+        sql_values = sql_values + '"' + user_is_lender_admin + '",'
 
-    user_is_suppliers_admin = request.args.get('user_is_suppliers_admin')
-    if user_is_suppliers_admin != '':
-        sql_parameters = sql_parameters + 'user_is_suppliers_admin,'
-        sql_values = sql_values + '"' + user_is_suppliers_admin + '",'
+    user_is_owner = request.args.get('user_is_owner')
+    if user_is_owner != '':
+        sql_parameters = sql_parameters + 'user_is_owner,'
+        sql_values = sql_values + '"' + user_is_owner + '",'
 
     db = Database()
 
-    if sql_parameters[len(sql_parameters)-1] == ',':
-        sql_parameters = sql_parameters[:len(sql_parameters)-1] + sql_parameters[(len(sql_parameters)-1+1):]
-    if sql_values[len(sql_values)-1] == ',':
-        sql_values = sql_values[:len(sql_values)-1] + sql_values[(len(sql_values)-1+1):]
+    if sql_parameters[len(sql_parameters) - 1] == ',':
+        sql_parameters = sql_parameters[:len(sql_parameters) - 1] + sql_parameters[(len(sql_parameters) - 1 + 1):]
+    if sql_values[len(sql_values) - 1] == ',':
+        sql_values = sql_values[:len(sql_values) - 1] + sql_values[(len(sql_values) - 1 + 1):]
 
     sql_string = sql_string + sql_parameters + ') VALUES ' + sql_values + ')'
     print(sql_string)
@@ -267,14 +286,20 @@ def new_user():
     c.close()
     return 'http200'
 
-@users.route('/deleteUser', methods=['GET'])
+
+@users.route('/users/delete', methods=['GET'])
+@login_required
 def delete_user():
+    user = flask_login.current_user
+    # only admin is allowed to delete a user in the user table
+    if not user.is_admin():
+        return 'http403'
+
     db = Database()
     ID = request.args.get('ID')
 
-    sql1 = 'DELETE FROM users WHERE ID = "'+ID+'"'
-    print(sql1)
-    sql2 = 'DELETE FROM pictures WHERE user_id = "'+ID+'"'
+    sql1 = 'DELETE FROM users WHERE ID = "' + ID + '"'
+    sql2 = 'DELETE FROM pictures WHERE user_id = "' + ID + '"'
 
     db.conn.row_factory = db.dict_factory
     c = db.conn.cursor()
@@ -282,17 +307,18 @@ def delete_user():
     c.execute(sql2)
     db.conn.commit()
 
-    directory = 'static/images/upload'
-    my_dir = os.path.join(basedir, directory)
-    for fname in os.listdir(my_dir):
-        if fname.startswith("-1"+ID):
-            os.remove(os.path.join(my_dir, fname))
+    pictures_dir = os.path.join(os.getcwd(), 'static\\images\\upload\\')
+    for filename in os.listdir(pictures_dir):
+        if filename.startswith("user_profile_picture_id_" + ID) or filename.startswith("user_register_upload_id_" + ID):
+            os.remove(os.path.join(pictures_dir, filename))
 
     c.close()
     return 'http200'
 
-@users.route('/saveUsersAsCSV')
-def save_equipment_as_csv():
+
+@users.route('/users/download-as-csv')
+@login_required
+def download_users_as_csv():
     db = Database()
     db.conn.row_factory = db.dict_factory
     c = db.conn.cursor()
@@ -314,12 +340,123 @@ def save_equipment_as_csv():
         # Writing data of CSV file
         csv_writer.writerow(eq.values())
 
-    return 'http200'
-
-@users.route('/downloadUsersCSV')
-def download_users_csv():
-    save_equipment_as_csv()
-
     with open("data_users.csv") as fp:
-        csv = fp.read()
-    return Response(csv,mimetype="text/csv",headers={"Content-disposition":"attachment; filename=data_users.csv"})
+        csv_file = fp.read()
+
+    user = flask_login.current_user
+
+    # admin is allowed to receive all the data
+    if user.is_admin():
+        return Response(csv_file, mimetype="text/csv",
+                        headers={"Content-disposition": "attachment; filename=data_users.csv"})
+    else:
+        return redirect(request.referrer)
+
+
+@users.route('/users/get-logged-in-users-data')
+@login_required
+def get_logged_in_users_data():
+    db = Database()
+
+    user_id = flask_login.current_user.get_id()
+
+    sql = 'SELECT ID, user_name, user_last_name, user_category, user_function, user_sex, user_email, user_telephone, ' \
+          'user_title, user_home_address, user_private_phone, user_is_pi, user_is_phd, user_in_date, user_out_date ' \
+          'FROM users WHERE ID = "' + str(user_id) + '" '
+    db.conn.row_factory = db.dict_factory
+    c = db.conn.cursor()
+    c.execute(sql)
+    return json.dumps(c.fetchall())
+
+
+@users.route('/users/get-by-id', methods=['GET'])
+@login_required
+def get_user_by_id():
+    db = Database()
+    user_id = request.args.get('ID')
+
+    sql = 'SELECT * FROM users WHERE ID = "' + user_id + '"'
+    db.conn.row_factory = db.dict_factory
+    c = db.conn.cursor()
+    c.execute(sql)
+    return json.dumps(c.fetchall())
+
+
+@users.route('/users/get-name-by-id', methods=['GET'])
+@login_required
+def get_user_name_by_id():
+    db = Database()
+    user_id = request.args.get('ID')
+
+    sql = 'SELECT ID, user_last_name, user_name FROM users WHERE ID = "' + user_id + '"'
+    db.conn.row_factory = db.dict_factory
+    c = db.conn.cursor()
+    c.execute(sql)
+    return json.dumps(c.fetchall())
+
+
+@users.route('/users/get-pop-up-info-by-id', methods=['GET'])
+@login_required
+def get_user_pop_up_info_by_id():
+    db = Database()
+    user_id = request.args.get('ID')
+
+    sql = 'SELECT ID, user_last_name, user_name, user_email, user_telephone FROM users WHERE ID = "' + user_id + '"'
+    db.conn.row_factory = db.dict_factory
+    c = db.conn.cursor()
+    c.execute(sql)
+    return json.dumps(c.fetchall())
+
+
+@users.route('/users/get-permissions-by-id', methods=['GET'])
+@login_required
+def get_user_permissions_by_id():
+    db = Database()
+    user_id = request.args.get('ID')
+
+    sql = 'SELECT ID, user_can_see_financial_data, user_is_admin, user_is_lender, user_is_lender_admin, ' \
+          'user_is_owner FROM users WHERE ID = "' + user_id + '" '
+    db.conn.row_factory = db.dict_factory
+    c = db.conn.cursor()
+    c.execute(sql)
+    return json.dumps(c.fetchall())
+
+
+@users.route('/users/get-by-access', methods=['GET'])
+@login_required
+def get_user_by_access():
+    db = Database()
+    access = request.args.get('access')
+
+    sql = 'SELECT ID, user_name, user_last_name, user_email FROM users WHERE user_is_allowed = "' + access + '"'
+    db.conn.row_factory = db.dict_factory
+    c = db.conn.cursor()
+    c.execute(sql)
+    return json.dumps(c.fetchall())
+
+
+@users.route('/users/get-owner-suggestion-by-name', methods=['GET'])
+@login_required
+def get_owner_suggestion_by_name():
+    db = Database()
+    name = request.args.get('name')
+
+    sql_string_last_name = 'SELECT ID, user_last_name, user_name FROM users WHERE user_last_name LIKE "%' + name + '%"'
+    db.conn.row_factory = db.dict_factory
+    c = db.conn.cursor()
+    c.execute(sql_string_last_name)
+    data_last_name = c.fetchall()
+
+    sql_string_first_name = 'SELECT ID, user_last_name, user_name FROM users WHERE user_name LIKE "%' + name + '%"'
+    c.execute(sql_string_first_name)
+    data_first_name = c.fetchall()
+
+    list_of_all_names = data_first_name + data_last_name
+
+    unique_users = []
+    for x in list_of_all_names:
+        if x not in unique_users:
+            unique_users.append(x)
+
+    c.close()
+    return json.dumps(unique_users)
